Conflicts:
        com/planet_ink/coffee_mud/Common/DefaultSession.java
	com/planet_ink/coffee_mud/Libraries/CMProtocols.java
	com/planet_ink/coffee_mud/Libraries/Socials.java
	com/planet_ink/coffee_mud/Libraries/interfaces/ProtocolLibrary.java
	com/planet_ink/coffee_mud/core/CMProps.java

closes toasted323/coffeemud-dedicated-heron#100

+++ com/planet_ink/coffee_mud/Common/FakeSession.java
@@ -773,7 +773,7 @@ public class FakeSession implements Session
 	}
 
 	@Override
-	public void initTelnetMode(final int mobbitmap)
+	public void initTelnetMode(final long mobbitmap)
 	{
 	}
