+++ com/planet_ink/coffee_mud/application/VFShell.java
@@ -1028,7 +1028,7 @@ public class VFShell
 					}
 
 					@Override
-					public void initTelnetMode(final int mobbitmap)
+					public void initTelnetMode(final long mobbitmap)
 					{
 					}
